+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 
 ## [7.5.1.0] (Unreleased)
 
+### Features
+
+- [usd#2435](https://github.com/Autodesk/arnold-usd/issues/2435) - Report USD warning and errors as Arnold warning and errors.
+
 ### Bug Fixes
 
 - [usd#2547](https://github.com/Autodesk/arnold-usd/issues/2547) - Changing material terminal is not updated with Hydra 2